Only generate occurrences when valid timing rule exists

+++ tasking/signals.py
@@ -12,11 +12,12 @@ from tasking.tools import generate_task_occurrences
 
 @receiver(post_save, sender='tasking.Task')
 # pylint: disable=unused-argument
-def create_user_profile(sender, instance, created, **kwargs):
+def create_occurrences(sender, instance, created, **kwargs):
     """
     Create occurrences when a task timing_rule changes
     """
-    # delete any existing occurrences
-    instance.taskoccurrence_set.all().delete()
-    # generate new occurrences
-    generate_task_occurrences(instance)
+    if instance.timing_rule:
+        # delete any existing occurrences
+        instance.taskoccurrence_set.all().delete()
+        # generate new occurrences
+        generate_task_occurrences(instance)
